dingies

+++ GameTest/src/com/example/gametest/Beat.java
@@ -1,8 +1,7 @@
 package com.example.gametest;
 
-
 public class Beat {
-	public double time;
+	public long time;
 	// intensity kan gebaseerd worden op amplitude, of later ook
 	// op de frequentiebanden waar ze in voorkomen bijvoorbeeld
 	public double intensity;
