+++ GameTest/src/com/example/gametest/Section.java
@@ -1,9 +1,8 @@
 package com.example.gametest;
 
-
 public class Section {
-	double startTime;
-	double endTime; // of length, kan ook
+	long startTime;
+	long endTime; // of length, kan ook
 	Beat[] beats;
 	double intensity;
 	double avgTempo;
